Rename files to fix python2 import error
